requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
Flask-MySQLdb==2.0.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
PyMySQL==1.1.0
Flask-Cors==4.0.0
cryptography
python-dotenv==1.1.1
mysql-connector-python
Werkzeug==3.1.3
